refactor: refactored favorites -- old way
