+++ frontend/src/components/FavoriteButton/FavoriteButton.tsx
@@ -1,5 +1,4 @@
-import { useBooks } from "components/Book";
-import React from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { HeartIconThin } from "../../assets/icons";
 import { StyledHeart } from "./FavoriteButton.styled";
 
@@ -9,8 +8,28 @@ interface IProp {
 }
 
 export const FavoriteButton: React.FC<IProp> = ({ itemId }) => {
-  const { favorites, addFavorite, removeFavorite } = useBooks();
+  const [favorites, setFavorites] = useState<string[]>([]);
+
+  useEffect(() => {
+    localStorage.setItem("favorites", JSON.stringify(favorites));
+  }, [favorites]);
+
+  useEffect(() => {
+    const savedFavorites = JSON.parse(
+      localStorage.getItem("favorites") || "[]"
+    );
+    setFavorites(savedFavorites);
+  }, []);
+
   const isFavorite = favorites.includes(itemId);
+
+  const addFavorite = useCallback((id: string) => {
+    setFavorites((prevFavorites) => [...prevFavorites, id]);
+  }, []);
+
+  const removeFavorite = useCallback((id: string) => {
+    setFavorites((prevFavorites) => prevFavorites.filter((fav) => fav !== id));
+  }, []);
 
   const toggleFavorite = () => {
     if (isFavorite) {
