+++ frontend/src/components/PageHeading/PageHeading.tsx
@@ -1,5 +1,14 @@
+import { BreadCrumbs, Label } from "pages/CommonPages.styled";
+import { useParams } from "react-router-dom";
+
 const PageHeading = () => {
-  return <></>;
+  const { category, subcategory, link } = useParams();
+  return (
+    <>
+      <BreadCrumbs>Каталог | {category} </BreadCrumbs>
+      <Label> {link || subcategory || category || "Усі книги"} </Label>
+    </>
+  );
 };
 
 export default PageHeading;
